# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Authentication Service API",
    "description": "Authentication, token issuance, and JWT validation for the Neosofia platform. Delegates user authentication to WorkOS; issues RS256-signed JWTs for platform-wide service-to-service and client authentication.",
    "version": "1.0.0",
    "contact": {
      "name": "Neosofia Platform Team"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8014",
      "description": "Local development"
    },
    {
      "url": "https://auth.neosofia.local",
      "description": "Production"
    }
  ],
...
  "paths": {
    "/api/token": {
      "post": {
        "tags": [
          "OAuth2 & Tokens"
        ],
        "summary": "Issue platform JWT",
        "description": "Issue RS256-signed platform JWT via OAuth2 implicit (session) or client_credentials (Basic auth) grant. Supports both human authentication (via sealed WorkOS session cookie) and service-to-service authentication (via service credentials).",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "oneOf": [
                  {
                    "type": "object",
                    "title": "Session Grant (Human)",
                    "description": "Implicit grant: no body required, token extracted from wos_session cookie"
                  },
                  {
                    "type": "object",
                    "title": "Client Credentials Grant (Service)",
                    "properties": {
                      "grant_type": {
                        "type": "string",
                        "enum": [
                          "client_credentials"
                        ],
                        "description": "Must be 'client_credentials'"
                      },
                      "audience": {
                        "type": "string",
                        "description": "Requested audience (service name) for the issued token"
                      }
                    },
                    "required": [
                      "grant_type"
                    ]
                  }
                ]
              }
            },
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "object",
                    "title": "Session Grant (Human)",
                    "description": "Implicit grant: no body required, token extracted from wos_session cookie"
                  },
                  {
                    "type": "object",
                    "title": "Client Credentials Grant (Service)",
                    "properties": {
                      "grant_type": {
                        "type": "string",
                        "enum": [
                          "client_credentials"
                        ],
                        "description": "Must be 'client_credentials'"
                      },
                      "audience": {
                        "type": "string",
                        "description": "Requested audience (service name) for the issued token"
                      },
                      "client_id": {
                        "type": "string",
                        "description": "Service client_id (alternative to Authorization: Basic)"
                      },
                      "client_secret": {
                        "type": "string",
                        "description": "Service client_secret (alternative to Authorization: Basic)"
                      }
                    },
                    "required": [
                      "grant_type"
                    ]
                  }
                ]
              }
            }
          }
        },
        "security": [
          {
            "BasicAuth": []
          },
          {
            "SessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Token issued successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (missing grant type, invalid credentials format)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized (invalid credentials, no session)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable (WorkOS, database, or config unavailable)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/token-inspect": {
      "get": {
        "tags": [
          "Debug"
        ],
        "summary": "Inspect platform JWT claims (debug, development only)",
        "description": "Decode a Bearer JWT and return its claims without validating issuer, audience, or signature. Only available when ENV=development or ENV=test. Returns 404 in all other environments. Rate-limited to 10 requests per minute per IP.",
        "security": [
          {
            "BearerToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Token valid, claims returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformJWTClaims"
                }
              }
            }
          },
          "400": {
            "description": "Malformed or unparseable JWT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (10 requests per minute per IP)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not available outside development/test environments",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-environments": [
          "development",
          "test"
        ],
        "x-debug-only": true
      }
    },
...
  },
  "components": {
    "schemas": {
      "TokenResponse": {
        "type": "object",
        "title": "Token Response",
        "description": "Successful token issuance response",
        "required": [
          "access_token",
          "token_type",
          "expires_in"
        ],
        "properties": {
          "access_token": {
            "type": "string",
            "description": "RS256-signed JWT token",
            "minLength": 50
          },
          "token_type": {
            "type": "string",
            "enum": [
              "Bearer"
            ],
            "description": "Always 'Bearer' per RFC 6750"
          },
          "expires_in": {
            "type": "integer",
            "minimum": 1,
            "description": "Token TTL in seconds"
          }
        }
      },
      "PlatformJWTClaims": {
        "type": "object",
        "title": "Platform JWT Claims",
        "description": "Standard and custom claims in platform JWT",
        "required": [
          "sub",
          "iat",
          "exp",
          "jti",
          "neosofia:token_type",
          "neosofia:token_version"
        ],
        "properties": {
          "sub": {
            "type": "string",
            "description": "Subject (user ID or service name)"
          },
          "iat": {
            "type": "integer",
            "description": "Issued at (Unix timestamp)"
          },
          "exp": {
            "type": "integer",
            "description": "Expiration time (Unix timestamp)"
          },
          "jti": {
            "type": "string",
            "description": "JWT ID (unique identifier)"
          },
          "azp": {
            "type": "string",
            "description": "Authorized party (RFC 7519); present on service tokens only, set to the client_id of the service credential"
          },
          "neosofia:token_type": {
            "type": "string",
            "enum": [
              "human",
              "service"
            ],
            "description": "Token principal class: 'human' for user sessions, 'service' for service-to-service credentials"
          },
          "neosofia:token_version": {
            "type": "integer",
            "minimum": 1,
            "description": "Claim schema version; consumers must reject tokens whose version they do not support"
          },
          "neosofia:roles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Array of role strings (e.g., ['care_team', 'supervisor'])"
          },
          "neosofia:tenant_uuid": {
            "type": "string",
            "description": "Optional: tenant/tenant ID (absent for service tokens)"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "title": "Error Response",
        "description": "Standard error response format",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Machine-readable error code (e.g., 'invalid_grant', 'unauthorized')"
          },
          "error_description": {
            "type": "string",
            "description": "Optional human-readable error details"
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "RS256-signed platform JWT in Authorization header"
      },
      "BasicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "Base64-encoded client_id:client_secret for service credentials grant"
      },
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "wos_session",
        "description": "Sealed WorkOS session cookie (HTTPOnly, Secure)"
      }
    }
  }
}